requirements.txt
asgiref==3.8.1
attrs==25.3.0
chardet==5.2.0
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
reportlab==4.4.1
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
